git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@400 49725315-0998-4498-bb4a-373bc1a52cb2

+++ LASI_Algorithm/SyntaciticAndSemanticStructures/UnderpinningTypes/Clause.cs
@@ -86,7 +86,7 @@ namespace LASI.Algorithm
         /// <summary>
         /// Gets the or sets the double precision numeric weight of the Clause.
         /// </summary>
-        public double Weight {
+        public decimal Weight {
             get;
             set;
         }
