+++ LASI_Algorithm/SyntaciticAndSemanticStructures/UnderpinningTypes/Phrase.cs
@@ -126,7 +126,7 @@ namespace LASI.Algorithm
         /// <summary>
         /// Gets the or sets the double precision numeric weight of the Phrase.
         /// </summary>
-        public double Weight {
+        public decimal Weight {
             get;
             set;
         }
